Struct token_store::Store

pub struct Store { /* fields omitted */ }

A token store

This struct allows you to store various values in a store and access them back using the provided tokens.

Methods

impl Store

pub fn new() -> Store

Create a new store

pub fn insert<V: Any + 'static>(&mut self, value: V) -> Token<V>

Insert a new value in this store

Returns a clonable token that you can later use to access this value.

pub fn get<V: Any + 'static>(&self, token: &Token<V>) -> &V

Access value previously inserted in this store

Panics if the provided token corresponds to a value that was removed.

pub fn get_mut<V: Any + 'static>(&mut self, token: &Token<V>) -> &mut V

Mutably access value previously inserted in this store

Panics if the provided token corresponds to a value that was removed.

pub fn remove<V: Any + 'static>(&mut self, token: Token<V>) -> V

Remove a value previously inserted in this store

Panics if the provided token corresponds to a value that was already removed.

pub fn with_value<V: Any + 'static, T, F>(
&mut self,
token: &Token<V>,
f: F
) -> T where
F: FnOnce(&mut StoreProxy, &mut V) -> T,

Create a sub-scope with access to a value

In the closure you provide, the value represented by token will be available as an argument, as well as a StoreProxy, which allows you to manipulate the other values of the store while this one is mutably borrowed.

Attempting to access again the same value from its token from within this closure is forbidden, and attempting to do so will result in a panic.

The StoreProxy provides the same access methods as the Store, including with_value, allowing you to create nested sub-scopes accessing multiple store values at the same time.

pub fn as_proxy<'a>(&'a mut self) -> StoreProxy<'a>

See this Store as a StoreProxy with no ongoing borrow

This can be usefull for code requiering access to a store, but wanting to be generic over being called from a value scope or not.

You can also use the From and Into traits to perform this conversion.